show rate cart/postponed

+++ src/main/java/com/example/MyBookShopApp/dto/response/BookDTO.java
@@ -8,7 +8,7 @@ public class BookDTO {
     private String title;
     private Short discount;
     private Byte isBestseller;
-    private Short rating; //количество звезд у книги
+    private Integer rating; //количество звезд у книги
     private String status; //KEPT (отложена), CART (в корзине) и PAID(куплена)
     private Integer price;
     private Integer discountPrice;
@@ -81,11 +81,11 @@ public class BookDTO {
         this.discount = discount;
     }
 
-    public Short getRating() {
+    public Integer getRating() {
         return rating;
     }
 
-    public void setRating(Short rating) {
+    public void setRating(Integer rating) {
         this.rating = rating;
     }
